ai-backend/requirements.txt
Flask==3.0.3
matplotlib==3.9.2
numpy==2.1.1
pandas==2.2.3
scikit_learn==1.5.2
tensorflow_intel==2.17.0
